# swagger/Trainingsplanner-postgres.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "laplanner",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/appointments/{id}": {
      "get": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_GetAppointmentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsAppointmentDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/appointments/{id}/modules": {
      "get": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_GetModulesByAppointmentId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsModuleDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{trainingsAppointmentId}": {
      "post": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_AddModuleToAppointment",
        "parameters": [
          {
            "name": "trainingsAppointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "moduleDtos",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TrainingsModuleDto"
                }
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsAppointmentTrainingsModuleDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{appointmentId}/modules/{moduleId}": {
      "delete": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_DeleteAppointmentWithModule",
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "202": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/trainingsexercises/modules/{trainingsModuleId}": {
      "get": {
        "tags": [
          "TrainingsExercise"
        ],
        "summary": "Returns the TrainingsExercise with the given Id",
        "operationId": "TrainingsExercise_ReadExercisesByModuleId",
        "parameters": [
          {
            "name": "trainingsModuleId",
            "in": "path",
            "required": true,
            "description": "Id of the required TrainingsModule",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "The TrainingsExercise",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsExerciseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/modules": {
      "post": {
        "tags": [
          "TrainingsModule"
        ],
        "summary": "Creates a new TrainingsModule from Body",
        "operationId": "TrainingsModule_CreateTrainingsModule",
        "requestBody": {
          "x-name": "trainingsModule",
          "description": "Id of the needed TrainingsModule",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsModuleDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "The TrainingsModule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsModuleDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TrainingsModule"
        ],
        "summary": "Updates the TrainingsModule with the given Id",
        "operationId": "TrainingsModule_UpdateTrainingsModule",
        "requestBody": {
          "x-name": "trainingsModule",
          "description": "The TrainingsModule with updated values (from HTTP-Body)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsModuleDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "The updated TrainingsModule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsModuleDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TrainingsModule"
        ],
        "summary": "Deletes the TrainingsModule with the given Id",
        "operationId": "TrainingsModule_DeleteTrainingsModule",
        "requestBody": {
          "x-name": "trainingsModule",
          "description": "The TrainingsModule with updated values (from HTTP-Body)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsModuleDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "204": {
            "description": "The deleted TrainingsModule"
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "TrainingsModule"
        ],
        "summary": "Returns all TrainingsModules",
        "operationId": "TrainingsModule_GetAllTrainingsModules",
        "responses": {
          "200": {
            "description": "All TrainingsModules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsModuleDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/modules/exercisemodule": {
      "post": {
        "tags": [
          "TrainingsModule"
        ],
        "summary": "Adds Exercise to a specific Module",
        "operationId": "TrainingsModule_AddExerciseToModule",
        "parameters": [
          {
            "name": "moduleId",
            "in": "query",
            "description": "Id of the needed TrainingsModule",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "exerciseId",
            "in": "query",
            "description": "Id of the needed Exercise",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "201": {
            "description": "TrainingsModuleTrainingsExerciseDto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsModuleTrainingsExerciseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/tagmodule": {
      "post": {
        "tags": [
          "TrainingsModule"
        ],
        "summary": "Adds Tag to a specific Module",
        "operationId": "TrainingsModule_AddTagToModule",
        "parameters": [
          {
            "name": "moduleId",
            "in": "query",
            "description": "Id of the needed TrainingsModule",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "tagId",
            "in": "query",
            "description": "the needed Exercise",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "201": {
            "description": "TrainingsModuleTrainingsExerciseDto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsModuleTrainingsExerciseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/modules/{moduleId}/tags/{tagId}": {
      "delete": {
        "tags": [
          "TrainingsModule"
        ],
        "summary": "Deletes the TrainingsModuleTag with the given Id from the given Module",
        "operationId": "TrainingsModule_DeleteTagByModuleId",
        "parameters": [
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "description": "The TrainingsModuleId (from HTTP-Body)",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "tagId",
            "in": "path",
            "required": true,
            "description": "The TagId(from HTTP-Body)",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "The deleted TrainingsModule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsModuleDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/modules/{id}/exercises": {
      "get": {
        "tags": [
          "TrainingsModule"
        ],
        "operationId": "TrainingsModule_GetAllExercisesByModuleId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsExerciseDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TrainingsAppointmentDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "trainingsGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsGroup": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsGroup"
              }
            ]
          },
          "trainingsAppointmentsTrainingsModules": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointmentTrainingsModule"
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}